// src/api/back/models/SubscriptionProductDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SubscriptionProductDto
 */
export interface SubscriptionProductDto {
    /**
     * 
     * @type {number}
     * @memberof SubscriptionProductDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof SubscriptionProductDto
     */
    months: number;
    /**
     * 
     * @type {number}
     * @memberof SubscriptionProductDto
     */
    pricePerMonth: number;
    /**
     * 
     * @type {number}
     * @memberof SubscriptionProductDto
     */
    discount: number;
}

export function SubscriptionProductDtoFromJSON(json: any): SubscriptionProductDto {
    return SubscriptionProductDtoFromJSONTyped(json, false);
}

export function SubscriptionProductDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubscriptionProductDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'months': json['months'],
        'pricePerMonth': json['pricePerMonth'],
        'discount': json['discount'],
    };
}

export function SubscriptionProductDtoToJSON(value?: SubscriptionProductDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'months': value.months,
        'pricePerMonth': value.pricePerMonth,
        'discount': value.discount,
    };
}